Refactoring: add TagProducer which manages tags
Collect responsibility of the stats tag management to the newly created
class, TagProducer.  A set of TagExtractor and a set of default Tag will
be handled via TagProducer class.

I'm also looking proper places to extract bootstrap config handling
around stats tag from Config::Utility. In this commit, leave the config
handling as it is for this time.

Signed-off-by: Taiki Ono <taiks.4559@gmail.com>